Lots of changes: - PixelFormat -> VideoFrameEncoding - Support for not being able to format a video frame - Fix all compiler warnings - renamed some internal BMDLCap vars - minor fixes - updated roadmap/todo/changelog

+++ src/microsoft_directshow/live_source_filter/CLiveSource.cpp
@@ -177,7 +177,7 @@ ULONG CLiveSource::GetMiscFlags()
 
 int CLiveSource::GetFrameQueueSize()
 {
-	return m_videoOutputPin->GetFrameQueueSize();
+	return (int)m_videoOutputPin->GetFrameQueueSize();
 }
